Updating samples folder path
