+++ samples/AspNet/Controllers/WeatherForecastController.cs
