+++ samples/AspNet/Global.asax.cs
