+++ samples/AspNet/Models/WeatherForecast.cs
