+++ samples/AspNet/OpenTelemetry.Exporter.AspNet.Win.csproj
@@ -83,23 +83,23 @@
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\OpenTelemetry.Api\OpenTelemetry.Api.csproj">
+    <ProjectReference Include="..\..\src\OpenTelemetry.Api\OpenTelemetry.Api.csproj">
       <Project>{99f8a331-05e9-45a5-89ba-4c54e825e5b2}</Project>
       <Name>OpenTelemetry.Api</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\..\src\OpenTelemetry.Instrumentation.AspNet.Win\OpenTelemetry.Instrumentation.AspNet.Win.csproj">
+    <ProjectReference Include="..\..\src\OpenTelemetry.Instrumentation.AspNet.Win\OpenTelemetry.Instrumentation.AspNet.Win.csproj">
       <Project>{b9eeacdd-cafa-4b75-a18d-898e7de21b17}</Project>
       <Name>OpenTelemetry.Instrumentation.AspNet.Win</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\..\src\OpenTelemetry.Instrumentation.Dependencies\OpenTelemetry.Instrumentation.Dependencies.csproj">
+    <ProjectReference Include="..\..\src\OpenTelemetry.Instrumentation.Dependencies\OpenTelemetry.Instrumentation.Dependencies.csproj">
       <Project>{d3ffbc59-2486-4f8f-bff1-fa95c84929e1}</Project>
       <Name>OpenTelemetry.Instrumentation.Dependencies</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\..\src\OpenTelemetry.Exporter.Jaeger\OpenTelemetry.Exporter.Jaeger.csproj">
+    <ProjectReference Include="..\..\src\OpenTelemetry.Exporter.Jaeger\OpenTelemetry.Exporter.Jaeger.csproj">
       <Project>{8d47e3cf-9ae3-42fe-9084-feb72d9ad769}</Project>
       <Name>OpenTelemetry.Exporter.Jaeger</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\..\src\OpenTelemetry\OpenTelemetry.csproj">
+    <ProjectReference Include="..\..\src\OpenTelemetry\OpenTelemetry.csproj">
       <Project>{ae3e3df5-4083-4c6e-a840-8271b0acde7e}</Project>
       <Name>OpenTelemetry</Name>
     </ProjectReference>
