+++ samples/AspNet/Properties/AssemblyInfo.cs
