+++ samples/AspNet/Views/Home/Index.cshtml
