+++ samples/AspNet/Views/Shared/_Layout.cshtml
