+++ samples/AspNet/Views/_ViewStart.cshtml
