+++ samples/Console/InstrumentationWithActivitySource.cs
