+++ samples/Console/Program.cs
