+++ samples/Console/TestConsoleExporter.cs
