+++ samples/Console/TestHttpClient.cs
