+++ samples/Console/TestJaegerExporter.cs
