+++ samples/Console/TestOTelShimWithConsoleExporter.cs
