+++ samples/Console/TestOtlpExporter.cs
