+++ samples/Console/TestPrometheusExporter.cs
