+++ samples/Console/TestRedis.cs
