+++ samples/Console/TestZPagesExporter.cs
