--- samples/Exporters/Web/appsettings.json
@@ -1,19 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "Jaeger": {
-    "ServiceName": "jaeger-test",
-    "Host": "localhost",
-    "Port": 6831
-  },
-  "Zipkin": {
-    "ServiceName": "zipkin-test",
-    "Endpoint": "http://localhost:9411/api/v2/spans"
-  }
-}
